Fix color mapping && Still Shift bug

+++ template.h
@@ -29,7 +29,7 @@ public:
     long long int computeArcDistance(int arc, int hue, int id);
     int computeNearRegion(int arc, int hue, int id);
     int targetHue(int arc, int hue, int id);
-    int counterOrcouterWise(int hue, int center);
+    float counterOrcouterWise(int hue, int center);
     int nearestDistance(int hue1, int hue2);
 };
 #endif // TEMPLATE_H
